requirements.txt
numpy==1.25.1
PyYAML==6.0
rdflib==7.0.0
Requests==2.31.0
scikit_learn==1.3.2
treelib==1.7.0